Add custom animations for dashboard cards and refactor navigation in login screen
