+++ lib/features/dashboard/ui/views/dashboard_screen.dart
@@ -1,13 +1,11 @@
-import 'package:az_task/core/animation/custom_fade_animation.dart';
 import 'package:az_task/core/common/widgets/spaces.dart';
 import 'package:az_task/core/common/widgets/theme_toggle_switch.dart';
 import 'package:az_task/core/images/assets_images.dart';
 import 'package:az_task/core/theme/app_colors.dart';
-import 'package:az_task/features/dashboard/data/card_model.dart';
 import 'package:az_task/features/dashboard/data/list_cards.dart';
 import 'package:az_task/features/dashboard/ui/widgets/custom_card.dart';
+import 'package:az_task/features/dashboard/ui/widgets/custom_grid_card.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_screenutil/flutter_screenutil.dart';
 
 class DashboardScreen extends StatelessWidget {
   const DashboardScreen({super.key});
@@ -45,6 +43,7 @@ class DashboardScreen extends StatelessWidget {
                 child: ListView.builder(
                   itemBuilder:
                       (context, index) => CustomCard(
+                        duration: (index + 2) * 180,
                         cardModel: cardList[index],
                         isRevresed: index % 2 == 0,
                       ),
@@ -71,91 +70,13 @@ class DashboardScreen extends StatelessWidget {
                           crossAxisCount: 2,
                         ),
                     itemBuilder:
-                        (_, index) => CustomGridCard(
-                          cardModel: cardList[index],
-                          isRevresed: index % 2 == 0,
-                        ),
+                        (_, index) =>
+                            CustomGridCard(cardModel: cardList[index]),
                     itemCount: cardList.length,
                   ),
                 ),
           );
         },
-      ),
-    );
-  }
-}
-
-class CustomGridCard extends StatelessWidget {
-  const CustomGridCard({
-    super.key,
-    required this.cardModel,
-    required this.isRevresed,
-  });
-  final CardModel cardModel;
-  final bool isRevresed;
-
-  @override
-  Widget build(BuildContext context) {
-    return CustomFadeInAnimation(
-      milliDuration: 600,
-      child: Card(
-        shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(24)),
-        color: Theme.of(context).cardColor,
-        child: Padding(
-          padding: const EdgeInsets.all(8.0),
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.center,
-            children: [
-              Text(
-                cardModel.title,
-                style: Theme.of(
-                  context,
-                ).textTheme.titleMedium?.copyWith(fontWeight: FontWeight.bold),
-              ),
-              verticalSpace(8),
-              Text(
-                cardModel.subtitle,
-                style: Theme.of(context).textTheme.bodyMedium,
-              ),
-              verticalSpace(16),
-              Text(
-                cardModel.count,
-                style: Theme.of(context).textTheme.titleMedium?.copyWith(
-                  fontWeight: FontWeight.bold,
-                  color: cardModel.iconColor,
-                  fontSize: 28,
-                ),
-              ),
-              verticalSpace(16),
-              Container(
-                alignment: Alignment.center,
-                padding: EdgeInsets.symmetric(horizontal: 20.w, vertical: 2.h),
-                decoration: BoxDecoration(
-                  color: AppColors.success.withAlpha(80),
-                  borderRadius: BorderRadius.circular(64),
-                ),
-                child: Center(
-                  child: Row(
-                    mainAxisAlignment: MainAxisAlignment.center,
-                    children: [
-                      Text(
-                        cardModel.percentage,
-                        style: Theme.of(context).textTheme.bodyMedium?.copyWith(
-                          fontWeight: FontWeight.bold,
-                          color: AppColors.success,
-                        ),
-                      ),
-                      Icon(
-                        Icons.arrow_outward_outlined,
-                        color: AppColors.success,
-                      ),
-                    ],
-                  ),
-                ),
-              ),
-            ],
-          ),
-        ),
       ),
     );
   }
